feat: cleaned many files by dividing them
