+++ src/engine/modules/ECSManager/LuaECSManager/MsgPackUtils.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <sol/sol.hpp>
+#include <msgpack.hpp>
+
+namespace rtypeEngine {
+    void serializeToMsgPack(const sol::object& obj, msgpack::packer<msgpack::sbuffer>& pk);
+    sol::object msgpackToLua(sol::state_view& lua, const msgpack::object& obj);
+}
